==> textovka/02_sparknlp.ipynb <==
import sys

from pyspark.sql import SparkSession
from pyspark.ml import PipelineModel

import sparknlp
from sparknlp.annotator import *
from sparknlp.common import *
from sparknlp.base import *
from sparknlp.pretrained import ResourceDownloader

spark = sparknlp.start()

print("Spark NLP version: ",sparknlp.version())
print("Spark version: ", spark.version)

import pandas as pd
predmety = pd.read_csv(r'C:\Users\zbyne\Documents\čokoláda\VŠE\datový projekt\Textová analytika - Projekt\BaseCleanUp\predmety_kit.csv')
df_exportedPandas = pd.read_csv(r'C:\Users\zbyne\Documents\čokoláda\VŠE\datový projekt\Textová analytika - Projekt\BaseCleanUp\Exported.csv')
df_exportedPandas = df_exportedPandas.iloc[:,[1,3,4,5,6]]
df_exportedPandas.head()

prdmt = list(predmety.iloc[:,2])

import numpy as np

spark.conf.set("spark.sql.execution.arrow.enabled", "true")
#create sparkdf from pandas dataframe
data = spark.createDataFrame(df_exportedPandas)


data.select("*").show()

document = DocumentAssembler() \
    .setInputCol('HODNOTA_CLOB_OCISTENO') \
    .setOutputCol('document') \
    .setCleanupMode("shrink")

language_detector = LanguageDetectorDL.pretrained("ld_wiki_tatoeba_cnn_21") \
    .setInputCols(["document"]) \
    .setOutputCol("LANG") \
    .setThreshold(0.8) \
    .setCoalesceSentences(True)

marian = MarianTransformer.pretrained("opus_mt_cs_en", "xx") \
    .setInputCols(["document"]) \
    .setOutputCol("translation")

sentence = SentenceDetector() \
    .setInputCols("translation") \
    .setOutputCol('sentence')

tokenizer = Tokenizer() \
    .setInputCols('sentence') \
    .setOutputCol('token')

stop_words = StopWordsCleaner.pretrained("stopwords_en", "en") \
        .setInputCols(["token"]) \
        .setOutputCol("cleanTokens")

lemmatizer = Lemmatizer() \
    .setInputCols(["cleanTokens"]) \
    .setOutputCol("lemma") \
    .setDictionary("./AntBNC_lemmas_ver_001.txt", value_delimiter ="\t", key_delimiter = "->")

pos = PerceptronModel.pretrained("pos_ud_ewt", "en") \
  .setInputCols(["translation", "token"]) \
  .setOutputCol("pos")

sentiment_detector = ViveknSentimentModel.pretrained() \
    .setInputCols(["translation", "token"]) \
    .setOutputCol("sentiment_result")

nlp_pipeline = Pipeline(stages=[document,language_detector,marian,sentence,tokenizer,stop_words, lemmatizer,pos,sentiment_detector])
model = nlp_pipeline.fit(data)
result = model.transform(data)

res = result.selectExpr('ID','OTAZKA','PEDAGOG','HODNOTA_CLOB_OCISTENO','LANG.result as LANG','token.result as token','lemma.result as lemma','pos.result as pos','sentiment_result.result as DetectSentiment','SENTIMENT')

type(res)

res_expPandasDF = res.toPandas()

res_expPandasDF.drop(columns = "Length",inplace = True)

res_expPandasDF.head()

#print(res_expPandasDF.iloc[1,4], + len(res_expPandasDF.iloc[1,4]))

listWithLength = []
for i in range(len(res_expPandasDF)):
    x = len(res_expPandasDF.iloc[i,4])
    listWithLength.append(x)

res_expPandasDF["Length"] = listWithLength

res_expPandasDF.to_csv(r'C:\Users\zbyne\Documents\čokoláda\VŠE\datový projekt\Textová analytika - Projekt\BaseCleanUp\EnglishLemmaCloud.csv',encoding="utf-8")

#ShrinkedDF.to_csv(r'C:\Users\zbyne\Documents\čokoláda\VŠE\datový projekt\Textová analytika - Projekt\BaseCleanUp\EnglishCloudWithLimitedLength.csv')
