==> textovka/05_finaltransformation.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv(r"D:\čokoláda\VŠE\Datový projekt\datový projekt\textovka\CleanedLemmaToTf-Idf.txt",engine = 'python')

df = df.iloc[:,[2,12,11]]

df["ConWord"] = np.nan
df["WordVal"] = np.nan

df.head()

df_positive=pd.read_csv(r'D:\čokoláda\VŠE\Datový projekt\datový projekt\textovka\Positive-Idf.txt',engine = 'python')
df_negative=pd.read_csv(r'D:\čokoláda\VŠE\Datový projekt\datový projekt\textovka\Negative-Idf.txt',engine = 'python')

df_positive["Sentiment"] = "Positive"
df_negative["Sentiment"] = "Negative"

df_importance = pd.concat([df_positive,df_negative])

df_importance.shape

df_importance.sort_values(by=["importance"],inplace = True,ascending = False)
df_importance.head()

print(df.dtypes)
df["ConWord"] = df["ConWord"].astype('object') #--> gotta be object otherwise can't store list

#for every row creates list of words if the word from df_importance is in
for i in range(len(df)):
    lstwrd = []
    
    for word in list(df_importance.iloc[:,0]):
        try:
            if word in df.iloc[i,1].split(" ",-1):
                lstwrd.append(word)
        except AttributeError:
            lstwrd = []

    if len(lstwrd) > 0:
        df.at[i,"ConWord"] = lstwrd #word
        
    else:
        df.at[i,"ConWord"] = []
        

#explore results
df.tail()

#just run some test
row_value = df.iloc[11,:].copy()
print(len(row_value["ConWord"]))
print(row_value["ConWord"])

#create copy of df(just to have some columns)
TestDf = df.iloc[0:2,:].copy()
#clear rows
TestDf = TestDf[0:0]
IterateDf = df.iloc[:,:].copy()

#for every row in iteratedf take that row and store it as row_value 
for row in range(len(IterateDf)):
    row_value = IterateDf.iloc[row,:].copy()

#for every word in row_value column con word
    for word in row_value["ConWord"]:
        #take one of the words and assign it as new value of that col
        row_value["ConWord"] = word
        TestDf = TestDf.append(row_value)
    

#Merge Df & df with importance
ResultDf = pd.merge(TestDf,df_importance, left_on = "ConWord", right_on = "word", how = "left",validate = "m:m")

ResultDf.head()

#Take only relevant cols
ResultDf = ResultDf[["ID","lemma_cleaned","Sentiment","word","importance"]]

#Save result
ResultDf.to_csv(r"D:\čokoláda\VŠE\Datový projekt\datový projekt\textovka\WordsWithImportance.csv")
